requirements.txt
absl-py==2.2.2
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
airportsdata==20250224
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
apache-tvm-ffi==0.1.5
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
autocommand==2.2.2
av==16.1.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
blake3==1.0.4
cachetools==5.5.2
cbor2==5.7.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
codetiming==1.4.0
colorama==0.4.6
compressed-tensors==0.9.2
contourpy==1.3.2
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cupy-cuda12x==13.4.1
cycler==0.12.1
datasets==3.6.0
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.17.0
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.124.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
fastrlock==0.8.3
filelock==3.18.0
flash_attn==2.7.4.post1
flashinfer-python==0.5.3
fonttools==4.58.2
frozenlist==1.6.0
fsspec==2025.3.0
gdown==5.2.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
hf-xet==1.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.0.0
inflect==7.3.1
interegular==0.3.3
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
loguru==0.7.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mathruler==0.1.0
matplotlib==3.10.3
mdurl==0.1.2
mistral_common==1.8.6
model-hosting-container-standards==0.1.11
modelscope==1.26.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
nanobind==2.7.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-frontend==1.16.0
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-cutlass-dsl==4.3.3
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==2.9.0
openai-harmony==0.0.8
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.9
orjson==3.10.18
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
pillow==11.2.1
pip==25.1
platformdirs==4.3.8
prometheus_client==0.22.0
prometheus-fastapi-instrumentator==7.1.0
propcache==0.3.1
protobuf==4.25.7
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pybase64==1.4.3
pycountry==24.6.1
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-extra-types==2.10.6
Pygments==2.19.1
pylatexenc==2.10
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.11
ray==2.52.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.25.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.28.0
setproctitle==1.3.6
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.50.0
style==1.1.0
supervisor==4.3.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensordict==0.8.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.4
tomli==2.3.0
torch==2.6.0
torchaudio==2.6.0
torchdata==0.11.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
typeguard==4.3.0
typer==0.15.4
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
update==0.0.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vllm==0.8.3
wandb==0.23.1
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.17
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
